IPAL52980: Wire Discretization with Table density fails   (StdMeshers_Regular_1D.cxx)
+ Compilation warnings

+ WIN specific errors
   (StdMeshers_Adaptive1D.cxx)

+++ src/SMESH/SMESH_MeshEditor.hxx
@@ -730,7 +730,7 @@ public:
   void sweepElement(const SMDS_MeshElement*                    elem,
                     const std::vector<TNodeOfNodeListMapItr> & newNodesItVec,
                     std::list<const SMDS_MeshElement*>&        newElems,
-                    const int                                  nbSteps,
+                    const size_t                               nbSteps,
                     SMESH_SequenceOfElemPtr&                   srcElements);
 
   /*!
